3. Django News Website/requirements.txt
jango==4.1.1
pillow==9.2.0